% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dribble.R
\name{is_dribble}
\alias{is_dribble}
\title{Is this a dribble?}
\usage{
is_dribble(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \link{dribble}, \code{FALSE} otherwise.
}
\description{
Is this a dribble?
}
\examples{
is_dribble(mtcars)
}
