% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protondrive-package.R
\docType{package}
\name{protondrive-package}
\alias{protondrive}
\alias{protondrive-package}
\title{protondrive: An Interface to Proton Drive}
\description{
Manage files on Proton Drive from R. Functions mirror the 'googledrive' package: list, find, upload, download, copy, move, rename, trash, and share files and folders. All end-to-end encryption, session handling, and API traffic are delegated to the official Proton Drive command-line interface, which is built on the Proton Drive SDK \url{https://github.com/ProtonDriveApps/sdk}. This is a third-party package that is not officially supported by Proton.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hinkelman.github.io/protondrive/}
  \item \url{https://github.com/hinkelman/protondrive}
  \item Report bugs at \url{https://github.com/hinkelman/protondrive/issues}
}

}
\author{
\strong{Maintainer}: Travis Hinkelman \email{thinkelman@pm.me} [copyright holder]

Authors:
\itemize{
  \item Travis Hinkelman \email{thinkelman@pm.me} [copyright holder]
}

}
\keyword{internal}
